initial support for s390x / ppc64le
This patch adds conditional compilation for code that is platform specific.
